Adds recipes page
Added a recipes page that where I only staged the actual recipe.html.
I am slightly confused right now and am hoping to check my work on
github to see how this pans out. I feel as if I have made an error and
this is a bad entry. Onward.